+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Added `Operator Version`, `Astarte Version`, `Health`, `Base API URL` and `Broker URL` columns as print columns to the Astarte CRD.
 - Added `API IP` and `Broker IP` columns as print columns to the ADI CRD.
+- Introduced a new API AstarteFDOIngress (v1alpha1) in the ingress.astarte-platform.org group to handle the ingress dedicated to Astarte FDO pairing requests.
 
 ## [26.5.1] - 2026-05-03